backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
Pillow
huggingface-hub
